* punt the NotePixmapPainter assert proble to Chris[1]
* replace the broken layout of ClefDialog with one that works and does an
  adequate job of restoring the original look

* Simplify BigArrowbutton, as it isn't necessary to go along specifying parents
  in Qt4 code, and new code should preferably be written without them.  These
  changes remove extra parameters, and simplify calling code where it occurs.

* Some initial work on some newly discovered style issues

[1] Start Rosegarden, draw a segment, go to a notation view and Segment->Add
Clef Change

With luck, you should see a clef dialog that has an empty staff in the middle,
although it may crash at this point.

If you get to the empty staff, click on one of the four arrow buttons, and you
will get a crash caused by aborting after a failed assert in NotePixmapPainter.h

[notation]  WARNING: NotePixmapPainter::begin(): Painter already active

rosegarden: src/gui/editors/notation/NotePixmapPainter.h:60: bool
Rosegarden::NotePixmapPainter::begin(QPaintDevice*): Assertion
`!m_myPainter.isActive()' failed.
Aborted (core dumped)

It's failing because the painter is already active, but the why of that and what
exactly this code is for (it has a strong air of cruft about it) I leave up to
Chris.

It may be related to something I've done setting up this dialog from
NotationView, but I don't *think* so.

Punt!


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10439 cf6a83e7-f017-4305-b0b3-678bde24f230