* fix recording, phase 1 (somewhat experimental)
1. Trace old Classic code to figure out that m_recording and everything that
affects it really was the  property controlled by the "Current? [x]" column in
the old device manager.  (There is no documentation for this anywhere, and the
name "isRecording()" could mean a lot of different things, but this
interpretation seems to be attested.)

2. Make the actual m_recording member variable a static const bool = TRUE!!! in
both places it occurs

3. Remove all accessor methods for setting this value

4. Remove all code that called the accessor methods

Now this property is TRUE, ALWAYS, and phase 2 will procede accordingly from
there after seeing what this does.  This commit is for convenient back tracking,
should it become necessary, and has only been tested to a successful link.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10458 cf6a83e7-f017-4305-b0b3-678bde24f230