Track header colour rework, phase 1:
* extend ColourType to include white notation and red notation, for track header
purposes; use this property, plus some extra kludge work outside of it

* extend PixmapFunctions to be able to recolour a pixmap using a specified,
rather than calculated saturation

* untold and forgotten bits and bobs that I hopefully documented adequately

NOTE: This was all meant to replace the "mask" code that Yves reported as a
hack.  The results of that have always left something to be desired, and this
was supposed to make it all look nice in addition to changing the colors around.

In practice, it seems to be a lot of work to move sideways, because these glyphs
are still rendered poorly.  I think I can fix that, and that's phase 2.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10552 cf6a83e7-f017-4305-b0b3-678bde24f230