Final Phase of white-notation-in-dialogs work
There's a lengthy comment in here somewhere about how I should refactor this and
clean it up now that I've worked the overall problem all the way through to the
end.

I'll do that another day.  I already wasted a great heaping lot of time
accomplishing something that wasn't really all that worthwhile.  The dialogs
looked crisp and sharp enough with black-on-white notation, but since I had done
this in the track headers, I decided to see what if, and what if turned out to
be so much work to bring about that I definitely intend to keep it now.

Oh well.  A waste of time on one level, and a fantastic and ambitious learning
experience on another.

Incidentally, I've looked at Rosegarden with no stylesheet, and decided not to
bother unhooking and rewiring any of the other hard coded stuff.  No stylesheet
gets you white on black track headers in a white field, and it might be worth
making that field black stylesheet or no, but the other hard coded bits seem
unobtrusive enough to just leave well enough alone.  Especially since everyone
is encouraged to run with the stylesheet anyway.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10577 cf6a83e7-f017-4305-b0b3-678bde24f230