* collapsing frames finally remember their settings
This introduces an interesting new wrinkle.  If you close the TPB frames down,
then quit and restart, the TPB won't expand to accommodate the new height of its
contents, and you wind up with completely illegible squished nothing.

Come to think of it, I think we may have alwasy had some variation on that
problem, but now it's orders of magnitude worse.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10617 cf6a83e7-f017-4305-b0b3-678bde24f230