New branch to go work on a very messy and experimental refactoring.
The aim is to pull out the laundry list of warning dialogs from the sequence
manager and replace them with a little warning tray on the status bar with one
or more status indicator icons.  (The most critical warnings, eg. JACK stopped
processing audio, will remain in the sequence manager.)

The work will be done in stages.  In the first stage, to be committed shortly, a
giant glob of dialog code gets commented out, and replaced with some bools and
accessor methods to provide a clean path for sending these status messages to
the GUI.

In the second stage, the WarningWidget will be written, hooked into this new
information pipeline, and have some preliminary proof of concept work done on
it.

If the work is worth pursuing further, the third stage will pull the warning
text over from the sequence manager and hook it up to the warning widget, so you
can find out what the trouble is.

The final stage will be to review the flags and accessor methods for
opportunites to refactor and clean up redundant or obsolete code.  For example,
while the accessors provide a clean and human-intelligible way to fetch these
pieces of status information, it is likely this could be done by tying directly
into the existing structure, instead of piggybacking on top of it.  This kind of
thing is more of a clean code vs. just git'r'done code issue.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/warning_widget@10619 cf6a83e7-f017-4305-b0b3-678bde24f230