Very credible if I do say so myself, with room to quibble over details.
For instance, should we show the icons when things are good, or when things are
bad?  Currently the objective is to get three icons to show up in a line and
live happily ever after, but it almost seems like the "nothing to see here"
state probably ought to be the exact opposite of that, actually.

Maybe I've even misunderstood what Chris suggested.  This used to use
green/yellow/red anyway.  Details, details, definitely nothing to lose sleep
over at this juncture.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/warning_widget@10639 cf6a83e7-f017-4305-b0b3-678bde24f230