Segno is visible now!
The font mapping business is another matter.  I have the idea, but where do the
numbers come from?  I think these were the numbers my tried and true and
familiar KCharSelect used to show, but the tool has been rendered UTTERLY
USELESS in KDE 4.2.

(To get the glyphs for drawing these icons, I finally had to resort to using
FontForge to copy and paste them out of the original font and into a new one
where I could find them.  This has all become a monumental pain in the ass
thanks to KCharSelect becoming UTTERLY USELESS. Let me scream that in all caps
again.  UTTERLY USELESS. WTF were they thinking? WTF is this pile of crap that
used to be a utility that was a daily staple of my work?  W. T. F. ?! )



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10667 cf6a83e7-f017-4305-b0b3-678bde24f230