* Added the WA_DeleteOnClose attribute to EditViewBase and some RG_DEBUG reports in destructors. * Crashes reliably when you close the matrix editor but only if you've already drawn a note.
The crash can be delayed by commenting out delete m_scene in ~MatrixWidget.
RG will then crash next time you open the segment and try to draw a note.

Looks to me like it's not getting through QT's child deletion process as comments in ~MatrixTool
are not showing but do if you comment out the delete of m_scene in ~MatrixWidget.

I don't have the QT source on this PC so gdb can't help in the location of the crash.
Hopefully someone with that installed and valgrind (or whatever) can help locate the problem.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10762 cf6a83e7-f017-4305-b0b3-678bde24f230