Confused/out of my depth.
Jani's valgrind o/p suggests my ControlRulerWidget::m_selectedElements list is the culprit for this crash:
==15885== Invalid free() / delete / delete[]
==15885==    at 0x4C260AD: operator delete(void*) (vg_replace_malloc.c:342)
==15885==    by 0x4C4CBB: Rosegarden::ViewElement::~ViewElement() (ViewElement.cpp:36)
==15885==    by 0x4C500C: Rosegarden::ViewElementList::~ViewElementList() (ViewElement.cpp:72)
==15885==    by 0x793E53: Rosegarden::ControlRulerWidget::~ControlRulerWidget() (ControlRulerWidget.cpp:58)

Presumably, because QT is trying to delete the elements contained in the list but they're already deleted?

This commit clears m_selectedElements during the ControlRulerWidget destructor and fixes the crash on my PC at least.
It doesn't feel right.

What don't I understand?

Is this the QT auto-cleanup mechanism? Should it be trying to cleanup a list of pointers?
If so, how am I supposed to say "don't clean these up, they belong to another widget"?


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10763 cf6a83e7-f017-4305-b0b3-678bde24f230