This is where rosegarden-project-packager will go to die, and emerge as real code.
We don't have any real code yet, but a general idea where it's going.

I think all the Perl hoopty blather was about parsing the XML to figure out what
needed included in the project package.  Well, why parse the XML when we have a
live RosegardenDocument to iterate through in the normal fashion?  I expect the
LilyPondExporter and similar code offers a substantially complete framework to
copy for iterating through all the segments on all the tracks in the document,
so it's just minor twiddling from there to extract the information we need.

I intend to remove the dependency on the external flac utility by switching over
to using libflac, if at all possible, and so on.  The overall goal here is to
reduce the number of weird command line dependencies to 0 if possible, while
adding new library dependencies as sparingly as possible.

We need to be able to:

  * import existing project packages

  * write project packages older versions of Rosegarden can at least partially
    understand

  * solve the "spaces in paths" bug that plagued the old project packager

Beyond that, the process for getting from one place to another may be completely
different in this new version, because nobody can make heads or tails of exactly
whatt the hell the original perl script in a bash script did.  It will be a lot
quicker and cleaner just to "reverse engineer" something that gets the job done.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/project_packager_rewrite@10773 cf6a83e7-f017-4305-b0b3-678bde24f230