Final version of the new trill with line indication
  * hide access to the trill with line mark, except in the Make Ornament dialog

  * add a new indication version that:

    i) fixes the internal rendering problem of not being able to have the trill
       line span tied notes (because it could never have a "scope" beyond the
       single note to which it was attached)

   ii) fixes the long-standing and irritating LilyPond export bug that prevented
       the trill and line from being used with tied notes for complicated
       technical reasons

  * add some new tests to marks-test.rg

  * edit my trumpet duet arrangement to make use of the new indications, and
    also breath marks

  * purge notation.rc of references to the old actions

NOTES: Most of the drawing code comes directly from the original code to draw
the mark version, but it had to be adapted, and there were some holes I plugged
with gobs of mud.  It's not very sophisticated, but I find the results
surprisingly acceptable.

The vertical layout code is not remotely sophisticated, but looking around, it
seems to be in good company there.  We have problems with things not moving out
of the way of ledger lines generally.  A battle for another day.

This is a bit touchy about placement for export purposes.  Best results seem to
be obtained when creating an indication that spans all but the last in a series
of tied notes; otherwise LilyPond wants to crowd the following note or rest too
much with the trill line.  Considering the number of other touchy indications we
have, I consider this quite acceptable.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10824 cf6a83e7-f017-4305-b0b3-678bde24f230