BAM!
Two hours of chasing pissy little oversights, and this finally made it through a
complete cycle.

- load zynfidel.rg
- export project as /tmp/failure.rgd
- import project from /tmp/failure.rgd

Booyah!

TODO:

The tmp working directory is not cleaned up.  I painted myself into a corner on
that one, and it needs some refactoring of passed parameters turned into member
variables or such so the cleanup stage has the right stuff.

Progress bars just sit there and don't even go into "knight rider" mode.  We
need to something clever with the progress bars in this one.  The cheap way out
works OK for LilyPond export most of the time, but some of these operations in
my test suite even on a fairly modest project file can be quite expensive on a
super fast computer.

Code could do with all kinds of cleanups and refactoring for consistency and
neatness.  All the std::string vs. QString every which way is a real mess, and
it was all done very haphazardly.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/project_packager_rewrite@10856 cf6a83e7-f017-4305-b0b3-678bde24f230