Calling it a day after marked improvements
DONE:

The panner to big wheel chaining seems to work well, and this all has the added
bonus that the wheel moves to show the feedback relationship between the two
acts.

When changing from big wheel zoom to axis-independent zoom and back, the zoom
resets itself.  This is mildly unsettling, but it feels like it may well be
workable, and it avoids problems trying to scale both axes at the same time
after they have moved out of proportion to each other.

Some improvement on the piano keyboard handling

NOT DONE:

It is still very possible to leave the keyboard mangled in one way or another,
and in the worst cases, hitting reset doesn't fix it yet.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10950 cf6a83e7-f017-4305-b0b3-678bde24f230