coming together now...
* percussion key button thingies work

* behavior seems sane in the absence of all that apply button/m_modified garbage
  I dumped with the previous commit (good riddance to an abysmally bad design
  feature)

I think I just have to sort the file load/save bits and make sure cut/copy/paste
and so on actually work and we can call it working.  The disabled thing is
unfortunate and difficult to deal with, because of how Qt wants to draw all
those disabled texts in this scheme.  I don't see a good way out of that one
yet.  It's pretty grizzly.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11011 cf6a83e7-f017-4305-b0b3-678bde24f230