throw instrument alias edit controls in IPB too
BUGS: first time I look at the first audio IPB, the label is completely blank
(and "REFRESH BUG!" is not showing, so it's somewhere else)

InputDialog has a parenting problem and comes up centered on the main window or
audio mixer.  I know we've had and solved this sort of thing before, but have no
idea how that was done off hand.  Someone else please feel free to fix that!

Connecting changed() to slotUpdateAllBoxes() wasn't enough to trigger the
desired refresh, and/or something else is busted there.  Changing the alias does
not trigger an immediate refresh of the IPB.

NOTE: the layout cheat seems valid enough. Qt will happily let you put more than
one widget at the same grid coordinates, and since the button is all the way
left (seemingly without making the whole layout wider, which is what I hoped
for) and the label centered, there's enough wiggle room to cheat this way and
avoid having to screw around with all the grid layout coordinates trying to get
the label to come out centered overall when anchored to a different cell.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11114 cf6a83e7-f017-4305-b0b3-678bde24f230