It should be possible to translate shortcuts now.  (The bulk of this commit was in data/ which got swallowed by the size limit on our list, and won't show up. Other developers might want to keep this in mind too.  It's a good idea to commit data/ separately if you are also including code changes other people might want to notice, because translation updates are massive and boring to look at, so we don't usually approve them when the list kicks them out.)
NOTES:  I went through yesterday and added a "note name" comment to all the
note-related strings "A" through "G" so it should be safe to assume that when
these strings appear in some other context, they are meant to be shortcuts.
This avoided having to do a lot of rewriting to the action file handling
whatnots to put a "shortcut" comment in.  Instead, I've put translator comments
into the generated .cpp file, which show up in Qt Linguist.

I haven't tried this to make sure the shortcuts are still working.  Everyone
should be on the watch for breakage here.

Finally, I translated the I, II, III garbage back to Do, Re, Mi in American and
British English.  I'm pretty sure that's how it used to go several years ago,
and we changed to I, II, III to be more international-friendly.  This should
make things less confusing for English speakers without confusing Romance
language speakers for whom Do, Re, Mi are absolute, rather than relative
pitches.  If this doesn't go over well, it can always be reverted.  I suppose it
might not go over well if a lot of non-English speakers are using the English
version anyway, for some reason.  Who knows.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11165 cf6a83e7-f017-4305-b0b3-678bde24f230