Try a new "UNSTABLE PRE-RELEASE" splash for development builds, including routine SVN builds.  The new make-release-tarball script will change this from UNSTABLE=1 to STABLE=1 as part of the "RELEASE" target.  All other targets will have the development splash, and that will be the normal state of any SVN tree as well.
I have an .xcf file of the splash itself, and don't necessarily suggest this is
actually a good concept in this form, but I think the idea in general has some
real merit.  The only downside is that I'll be looking at the ugly splash
perpetually, as will many other core users who always run development builds day
to day.  Oh well.  Let's try it anyway.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11265 cf6a83e7-f017-4305-b0b3-678bde24f230