* begin restoring the cursor [blah] and select actions, which are proving to be   viciously nasty
* add the "[blah] bar" variants back in, which seem to have disappeared from our
  .rc files somewhere along the way, or missed getting moved from somewhere else

* icon tweaks for the new actions

* remove obsolete MatrixStaff class to attic/ and purge references

We're just about ready to wipe out attic/ I hope.  I THINK these are the final
frontier here.  If not, it may be fair to say anything else that's still missing
is so obscure it's OK to leave it rotting, because I didn't find it after a very
thorough search for missing functionality.

Working code to follow whenever I manage to get it working, with notation to
follow after that.  This is all a puzzler, and so far it looks like the two
implementations are going to have to be quite different from each other.

(This should all compile.  I made clean and did a test build.  The offending
code is commented out at the moment. If anyone finds that not to be the case,
let me know.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11336 cf6a83e7-f017-4305-b0b3-678bde24f230