Go ahead, Thorsten.  It's OK to scream and throw a brick at my head.
I'm a bad man, a bad bad man.

But this is one last minute thing that really was worth doing, and it was good
timing that user happened to precipitate all of this today and put this in
motion.  This is a seriously needed improvement, because I'm living proof that
users won't manage their audio files sensibly if we leave them to their own
devices, and provide them no better tools than we traditionally have.

(For those who don't want to bother reading the code, if you try to record audio
on "Untitled" now, it forces you to read an informative dialog box, then save
your file as something.  If you cancel, you record anyway, but at least you were
warned.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11345 cf6a83e7-f017-4305-b0b3-678bde24f230